all aspects of POST to /projects are working including adding info to db and creating github repo and associated project board with three columns
