+++ server/api/projects.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const createToken = require('./utils')
+const {createToken, githubRepoAndProjectBoardCreation} = require('./utils')
 const octokit = require('@octokit/rest')()
 const { Project, Collaboration, Repo } = require('../db/models')
 module.exports = router
@@ -30,60 +30,60 @@ router.get('/:projectId', (req, res, next) => {
   .catch(next)
 })
 
-function githubRepoAndProjectBoardCreation (name, description) {
-  let project_id;
+// function githubRepoAndProjectBoardCreation (name, description) {
+//   let project_id;
 
-  let columnIds = {
-    toDoColumnId: null,
-    inProgressColumnId: null,
-    doneColumnId: null
-  }
+//   let columnIds = {
+//     toDoColumnId: null,
+//     inProgressColumnId: null,
+//     doneColumnId: null
+//   }
 
-  return octokit.repos.createForOrg({
-    headers,
-    org: 'Code-Bono-Projects',
-    name,
-    description,
-    has_projects: true
-  })
-  .then((repo) => {
-    const repoNameForProjectBoard = repo.data.name
-    return octokit.projects.createRepoProject({
-      headers,
-      owner: 'Code-Bono-Projects',
-      repo: repoNameForProjectBoard,
-      name: repoNameForProjectBoard
-    })
-  })
-  .then(projectBoard => {
-    project_id = projectBoard.data.id
-    return octokit.projects.createProjectColumn({
-      headers,
-      project_id,
-      name: 'To Do'
-    })
-  })
-  .then(toDoColumn => {
-    columnIds.toDoColumnId = toDoColumn.data.id
-    return octokit.projects.createProjectColumn({
-      headers,
-      project_id,
-      name: 'In Progress'
-    })
-  })
-  .then(inProgressColumn => {
-    columnIds.inProgressColumnId = inProgressColumn.data.id
-    return octokit.projects.createProjectColumn({
-      headers,
-      project_id,
-      name: 'Done'
-    })
-  })
-  .then(doneColumn => {
-    columnIds.doneColumnId = doneColumn.data.id
-    return columnIds
-  })
-}
+//   return octokit.repos.createForOrg({
+//     headers,
+//     org: 'Code-Bono-Projects',
+//     name,
+//     description,
+//     has_projects: true
+//   })
+//   .then((repo) => {
+//     const repoNameForProjectBoard = repo.data.name
+//     return octokit.projects.createRepoProject({
+//       headers,
+//       owner: 'Code-Bono-Projects',
+//       repo: repoNameForProjectBoard,
+//       name: repoNameForProjectBoard
+//     })
+//   })
+//   .then(projectBoard => {
+//     project_id = projectBoard.data.id
+//     return octokit.projects.createProjectColumn({
+//       headers,
+//       project_id,
+//       name: 'To Do'
+//     })
+//   })
+//   .then(toDoColumn => {
+//     columnIds.toDoColumnId = toDoColumn.data.id
+//     return octokit.projects.createProjectColumn({
+//       headers,
+//       project_id,
+//       name: 'In Progress'
+//     })
+//   })
+//   .then(inProgressColumn => {
+//     columnIds.inProgressColumnId = inProgressColumn.data.id
+//     return octokit.projects.createProjectColumn({
+//       headers,
+//       project_id,
+//       name: 'Done'
+//     })
+//   })
+//   .then(doneColumn => {
+//     columnIds.doneColumnId = doneColumn.data.id
+//     return columnIds
+//   })
+// }
 
 router.post('/', (req, res, next) => {
   const userId = req.body.userId
@@ -91,10 +91,11 @@ router.post('/', (req, res, next) => {
   const name = req.body.proposalName
   const description = req.body.proposalDescription
   const repoName = name.split(' ').join('-')
+  let repoId;
+
   Project.findOrCreate({
     where: {
       name,
-      repoName,
       description,
       proposalId
     }
@@ -102,21 +103,28 @@ router.post('/', (req, res, next) => {
   .tap(([project, created]) => {
     if(created) {
       return githubRepoAndProjectBoardCreation(repoName, description)
-      .then((stuff) => {
-        console.log('STUFF from github', stuff)
-        // return Repo.create({
-        //   name: repoName,
-
-        // })
+      .then((githubProjectColumns) => {
+        const {toDoColumnId, inProgressColumnId, doneColumnId} = githubProjectColumns
+        return Repo.create({
+          name: repoName,
+          toDoColumnId,
+          inProgressColumnId,
+          doneColumnId
+        })
+        .then(createdRepo => {
+          repoId = createdRepo.dataValues.id
+        })
+        .catch(next)
       })
-
     }
   })
   .spread((project, created) => {
+    console.log('createdRepo???', repoId)
+    project.setRepo(repoId)
     return project.addUsers(userId)
   })
-  .then(projectAndUser => {
-    res.status(201).json(projectAndUser)
+  .then(() => {
+    res.sendStatus(201)
   })
   .catch(next)
 })
