+++ server/db/models/project.js
@@ -6,18 +6,6 @@ const Project = db.define('project', {
     type: Sequelize.STRING,
     unique: true
   },
-  repoName: {
-    type: Sequelize.STRING,
-    unique: true
-  },
-  repoURL: {
-    type: Sequelize.VIRTUAL,
-    get() {
-      return (
-        `https://github.com/Code-Bono-Projects/${this.getDataValue('repoName')}`
-      );
-    }
-  },
   description: {
     type: Sequelize.TEXT
   },
